feat: 화면 연결 및 리팩토링 1차 완료
#31
