+++ app/src/main/res/layout/fragment_cart.xml
@@ -10,9 +10,9 @@
     tools:context=".CartFragment">
 
 
-                <androidx.recyclerview.widget.RecyclerView
-                    android:id="@+id/recyclercart_item"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent" />
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recyclercart_item"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 
 </LinearLayout>
